More robust helper empty file and dir creation.
Ensure empty file and dir have for sure permissions 0000.  Ensure the
file is actually empty. For a dir, do not make such check (because it
would require changing permissions first), but the empty dir is always
mounted read only, so it cannot for sure be entered in jail (it
cannot be chmoded to be listable).