Extend support for configuring ports access
It is now possible to configure open ports via config files, not only
command line. Pasta config syntax is supported, in particular "auto"
keyword that makes all ports that are open in the jail accessible from
the host. Publishing ports from host to jail is supported. UDP is
supported in addition to TCP.