More integration tests for exposing ports
Changes the wait in which tests wait for netcat server to become
available. Uses netcat verbose output to determine when port is bound,
instead of trying to connect several time with delays. Trying to
connect several times resulted in long test execution times for tests
that expect port to be not exposed.